dj3test/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.0.8
django-celery-results==1.2.1
django-debug-toolbar==2.2
django-extensions==3.0.3
django-filter==2.3.0
djangorestframework==3.11.0
Faker==4.1.1
kombu==4.6.11
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
vine==1.3.0
